requirements.txt
docopt
python-msgpack
Flask
avocado>=2.4,<2.5
Django>=1.6,<1.7
